Enhanced and Improved FX Parser Current Capabilities:
Now capable of parsing all rendering state groups.
States can be added to passes.
Pending Improvements:

Need to finalize the capability to set a whole group of states within a pass, as defined somewhere in the shader.
Need to complete the application of states for state descriptors. Currently, the blending state is fully operational, and the depth state is partially implemented.