The parser has been significantly refactored.
At first I iteratively got rid of grammar ambiguities - there were a lot of shift/reduce conflicts.

Now there is only one conflict left between function_declaration and var_decl. This will need to be sorted out.

Then there was an attempt to force the parser to return values from the rules, and not directly modify the effect object, so that in the future, I could beautifully build an AST when I get to parsing expressions and function bodies...